requirements.txt

airbyte>=0.25.0
prefect>=2.14.0
great_expectations>=0.17.0
pandas>=2.0.0
numpy>=1.24.0
snowflake-connector-python[pandas]>=3.0.0
